+++ eloc610LowPowerPartition/include/project_config.h
@@ -24,7 +24,7 @@
 
         #define BLUETOOTH_CLASSIC
 
-        #define VERSION "ELOC-P_V1.55"
+        #define VERSION "ELOC-P_V1.56"
 
         #define STATUS_LED          GPIO_NUM_4
         #define BATTERY_LED         GPIO_NUM_4